requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
cssselect==1.1.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.5
django-allauth==0.44.0
django-markdownx==3.0.1
django-taggit==1.3.0
idna==2.10
lxml==4.6.2
Markdown==3.3.3
markdown2==2.4.0
oauthlib==3.1.0
Pillow==8.1.0
pycparser==2.20
Pygments==2.7.4
PyJWT==2.0.1
pyquery==1.4.3
python-decouple==3.4
python3-openid==3.2.0
pytz==2020.5
readtime==1.1.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
